AP_NavEKF2: split otp flow from PosVelNED
